requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
django-background-tasks
django-cors-headers
requests
python-dotenv
rest_framework
gunicorn
psycopg2-binary
dj-database-url
whitenoise
django-storages 
boto3
